refactor(fe): 라우터 전체적으로 수정, 디자인 일관화
