+++ frontend/src/app/components/WelcomePage.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { motion } from "framer-motion";
-import { Button } from "./ui/button";
+import { Button } from "@/app/components/ui/button";
 import { useRouter } from "next/navigation";
 
 export function WelcomePage() {
