--- frontend/src/app/components/notice/EditNoticeModal.tsx
@@ -1,101 +1,0 @@
-'use client';
-
-import { useState, useEffect } from 'react';
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../ui/dialog";
-import { Button } from "../ui/button";
-import { Input } from "../ui/input";
-import { Label } from "../ui/label";
-import { Textarea } from "../ui/textarea";
-import { apiFetch } from '../../../lib/backend/client';
-
-interface EditNoticeModalProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  onSuccess: () => void;
-  noticeId: number;
-  initialData: {
-    title: string;
-    content: string;
-  };
-}
-
-export function EditNoticeModal({ open, onOpenChange, onSuccess, noticeId, initialData }: EditNoticeModalProps) {
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    if (open && initialData) {
-      setTitle(initialData.title);
-      setContent(initialData.content);
-    }
-  }, [open, initialData]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-
-    try {
-      await apiFetch(`/api/v1/notices/${noticeId}`, {
-        method: 'PUT',
-        body: JSON.stringify({
-          title,
-          content,
-        }),
-      });
-
-      onSuccess();
-      onOpenChange(false);
-    } catch (error) {
-      console.error('공지사항 수정 실패:', error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-[425px]">
-        <DialogHeader>
-          <DialogTitle>공지사항 수정</DialogTitle>
-        </DialogHeader>
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div className="space-y-2">
-            <Label htmlFor="title">제목</Label>
-            <Input
-              id="title"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="공지사항 제목을 입력하세요"
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="content">내용</Label>
-            <Textarea
-              id="content"
-              value={content}
-              onChange={(e) => setContent(e.target.value)}
-              placeholder="공지사항 내용을 입력하세요"
-              rows={6}
-              required
-            />
-          </div>
-          <div className="flex justify-end space-x-2">
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => onOpenChange(false)}
-              disabled={isLoading}
-            >
-              취소
-            </Button>
-            <Button type="submit" disabled={isLoading}>
-              {isLoading ? '수정 중...' : '수정'}
-            </Button>
-          </div>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-}
