+++ frontend/src/app/mypage/notices/CreateNoticeModal.tsx
@@ -1,0 +1,91 @@
+'use client';
+
+import { useState } from 'react';
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/app/components/ui/dialog";
+import { Button } from "@/app/components/ui/button";
+import { Input } from "@/app/components/ui/input";
+import { Label } from "@/app/components/ui/label";
+import { Textarea } from "@/app/components/ui/textarea";
+import { apiFetch } from "@/lib/backend/client";
+
+interface CreateNoticeModalProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  onSuccess: () => void;
+}
+
+export function CreateNoticeModal({ open, onOpenChange, onSuccess }: CreateNoticeModalProps) {
+  const [title, setTitle] = useState('');
+  const [content, setContent] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsLoading(true);
+
+    try {
+      await apiFetch('/api/v1/notices', {
+        method: 'POST',
+        body: JSON.stringify({
+          title,
+          content,
+        }),
+      });
+
+      setTitle('');
+      setContent('');
+      onSuccess();
+      onOpenChange(false);
+    } catch (error) {
+      console.error('공지사항 생성 실패:', error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent className="sm:max-w-[425px]">
+        <DialogHeader>
+          <DialogTitle>공지사항 작성</DialogTitle>
+        </DialogHeader>
+        <form onSubmit={handleSubmit} className="space-y-4">
+          <div className="space-y-2">
+            <Label htmlFor="title">제목</Label>
+            <Input
+              id="title"
+              value={title}
+              onChange={(e) => setTitle(e.target.value)}
+              placeholder="공지사항 제목을 입력하세요"
+              required
+            />
+          </div>
+          <div className="space-y-2">
+            <Label htmlFor="content">내용</Label>
+            <Textarea
+              id="content"
+              value={content}
+              onChange={(e) => setContent(e.target.value)}
+              placeholder="공지사항 내용을 입력하세요"
+              rows={6}
+              required
+            />
+          </div>
+          <div className="flex justify-end space-x-2">
+            <Button
+              type="button"
+              variant="outline"
+              onClick={() => onOpenChange(false)}
+              disabled={isLoading}
+            >
+              취소
+            </Button>
+            <Button type="submit" disabled={isLoading}>
+              {isLoading ? '작성 중...' : '작성'}
+            </Button>
+          </div>
+        </form>
+      </DialogContent>
+    </Dialog>
+  );
+}
